homework01/期中/product-search-app/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
elasticsearch==8.10.0
python-dotenv==1.0.0